Add script to generate csv for 1 million entries
Currently needs to be run 10 times to generate 10 million entries, so there is room for optimization
