+++ package.json
@@ -32,6 +32,7 @@
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.24.1",
     "css-loader": "^1.0.1",
+    "csv-write-stream": "^2.0.0",
     "enzyme": "^3.7.0",
     "enzyme-adapter-react-16": "^1.6.0",
     "enzyme-to-json": "^3.3.4",
@@ -43,6 +44,7 @@
     "react-test-renderer": "^16.6.0",
     "regenerator-runtime": "^0.12.1",
     "style-loader": "^0.23.1",
+    "through2": "^3.0.0",
     "webpack": "^2.2.1",
     "webpack-cli": "^3.1.2"
   },
